+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0 - 2019-10-22
+
+* Working & tested in a client's website
+
+## 0.5.1 - 2019-10-07
+
+* Fixed Controller class name
+
+## 0.5.0 - 2019-10-07
+
+* Added Controller class
+
 ## 0.4.0 - 2019-10-04
 
 * All variables outputting